+++ app/src/main/java/jp/chau2chaun2/honkot/sample/multimodule/di/FragmentModule.kt
@@ -3,7 +3,7 @@ package jp.chau2chaun2.honkot.sample.multimodule.di
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
 import jp.chau2chaun2.honkot.sample.multimodule.ui.fragment.ListQiitaDocFragment
-import jp.chau2chaun2.honkot.sample.multimodule.ui.fragment.ListDatabaseFragment
+import jp.chau2chaun2.honkot.sample.multimodule.feature.list_database.ListDatabaseFragment
 import jp.chau2chaun2.honkot.sample.multimodule.ui.fragment.SelectFragment
 
 @Module
@@ -16,5 +16,5 @@ abstract class FragmentModule {
     abstract fun contributeListAPIFragment(): ListQiitaDocFragment
 
     @ContributesAndroidInjector
-    abstract fun contributeListDatabaseFragment(): ListDatabaseFragment
+    abstract fun contributeListDatabaseFragment(): jp.chau2chaun2.honkot.sample.multimodule.feature.list_database.ListDatabaseFragment
 }
