+++ app/src/main/java/jp/chau2chaun2/honkot/sample/multimodule/di/ViewModelModule.kt
@@ -7,7 +7,7 @@ import dagger.Binds
 import dagger.Module
 import dagger.multibindings.IntoMap
 import jp.chau2chaun2.honkot.sample.multimodule.CustomApplication
-import jp.chau2chaun2.honkot.sample.multimodule.vm.ListDatabaseFragmentViewModel
+import jp.chau2chaun2.honkot.sample.multimodule.feature.list_database.ListDatabaseFragmentViewModel
 import jp.chau2chaun2.honkot.sample.multimodule.vm.ListQiitaDocFragmentViewModel
 
 @Module
@@ -26,6 +26,6 @@ interface ViewModelModule {
 
     @Binds
     @IntoMap
-    @ViewModelKey(ListDatabaseFragmentViewModel::class)
-    fun listDatabaseFragmentViewModel(viewModel: ListDatabaseFragmentViewModel): ViewModel
+    @ViewModelKey(jp.chau2chaun2.honkot.sample.multimodule.feature.list_database.ListDatabaseFragmentViewModel::class)
+    fun listDatabaseFragmentViewModel(viewModel: jp.chau2chaun2.honkot.sample.multimodule.feature.list_database.ListDatabaseFragmentViewModel): ViewModel
 }
