+++ feature/list_database/build.gradle
@@ -1,23 +1,20 @@
-apply plugin: 'com.android.application'
+apply plugin: 'com.android.library'
 apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-android-extensions'
 apply plugin: 'kotlin-kapt'
-apply plugin: "androidx.navigation.safeargs.kotlin"
+apply plugin: 'kotlin-android-extensions'
 
 android {
     compileSdkVersion 29
     buildToolsVersion "29.0.2"
 
     defaultConfig {
-        applicationId "jp.chau2chaun2.honkot.sample.multimodule"
         minSdkVersion 21
         targetSdkVersion 29
         versionCode 1
         versionName "1.0"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-
-        vectorDrawables.useSupportLibrary = true
+        consumerProguardFiles 'consumer-rules.pro'
     }
 
     buildTypes {
@@ -45,49 +42,20 @@ android {
     }
 }
 
-kapt {
-    // Avoid NonExistentClass Error
-    correctErrorTypes = true
-}
-
 dependencies {
-    implementation project(":data:api")
     implementation project(":data:model")
-    implementation project(":data:dao")
     implementation project(":data:repository")
     implementation project(":util")
-    implementation project(":feature:list_database")
 
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
     implementation 'androidx.appcompat:appcompat:1.1.0'
-    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
-    testImplementation 'junit:junit:4.13'
+    implementation 'androidx.core:core-ktx:1.2.0'
+    testImplementation 'junit:junit:4.12'
     androidTestImplementation 'androidx.test.ext:junit:1.1.1'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
 
-    // kotlin
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    implementation "org.jetbrains.kotlin:kotlin-reflect:$kotlin_version"
-    kapt "androidx.databinding:databinding-compiler:3.6.2"
-
-    // Network Util
-    debugImplementation 'com.facebook.stetho:stetho:1.5.1'
-    debugImplementation 'com.facebook.stetho:stetho-okhttp3:1.5.1'
-    debugImplementation 'com.facebook.stetho:stetho-urlconnection:1.5.1'
-    implementation 'com.squareup.retrofit2:retrofit:2.8.1'
-    implementation 'com.squareup.retrofit2:converter-gson:2.8.1'
-    implementation 'com.squareup.okhttp3:logging-interceptor:4.5.0'
-
-    // Database Util (Orma)
-    annotationProcessor "com.github.maskarade.android.orma:orma-processor:6.0.2"
-    implementation "com.github.maskarade.android.orma:orma:6.0.2"
-    kapt "com.github.maskarade.android.orma:orma-processor:6.0.2"
-
-    // RxJava/RxAndroid/RxKotlin
-    implementation 'io.reactivex.rxjava2:rxandroid:2.1.1'
-    implementation 'io.reactivex.rxjava2:rxjava:2.2.19'
-    implementation 'io.reactivex.rxjava2:rxkotlin:2.4.0'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
 
     // Dagger
     implementation "com.google.dagger:dagger:2.27"
@@ -106,10 +74,4 @@ dependencies {
 
     // CircleImageView
     implementation 'de.hdodenhof:circleimageview:2.0.0'
-
-    // For Picasso to download and show image from web
-    implementation 'com.squareup.picasso:picasso:2.5.2'
-
-    // threetenabp
-    implementation 'com.jakewharton.threetenabp:threetenabp:1.2.3'
 }
